Move scalatestcontrib to the `shared` module
