+++ casper/src/test/scala/coop/rchain/casper/MergeabilityRules.scala
@@ -7,7 +7,7 @@ import cats.implicits._
 import coop.rchain.casper.helper.HashSetCasperTestNode
 import coop.rchain.casper.helper.HashSetCasperTestNode._
 import coop.rchain.casper.protocol.DeployData
-import coop.rchain.casper.scalatestcontrib._
+import coop.rchain.shared.scalatestcontrib._
 import coop.rchain.casper.util.ConstructDeploy
 import coop.rchain.casper.util.GenesisBuilder.GenesisContext
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
