+++ casper/src/test/scala/coop/rchain/casper/MultiParentCasperSmokeSpec.scala
@@ -3,7 +3,7 @@ package coop.rchain.casper
 import cats.implicits._
 import coop.rchain.casper.helper.HashSetCasperTestNode
 import coop.rchain.casper.helper.HashSetCasperTestNode._
-import coop.rchain.casper.scalatestcontrib._
+import coop.rchain.shared.scalatestcontrib._
 import coop.rchain.casper.util.ConstructDeploy
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
 import monix.execution.Scheduler.Implicits.global
