+++ casper/src/test/scala/coop/rchain/casper/api/ListeningNameAPITest.scala
@@ -4,7 +4,7 @@ import cats.implicits._
 import coop.rchain.casper.helper.HashSetCasperTestNode
 import coop.rchain.casper.helper.HashSetCasperTestNode._
 import coop.rchain.casper.protocol._
-import coop.rchain.casper.scalatestcontrib._
+import coop.rchain.shared.scalatestcontrib._
 import coop.rchain.casper.util.{ConstructDeploy, GenesisBuilder}
 import coop.rchain.models.Expr.ExprInstance.GInt
 import coop.rchain.models._
