+++ shared/src/test/scala/coop/rchain/shared/scalatestcontrib.scala
@@ -1,9 +1,9 @@
-package coop.rchain.casper
+package coop.rchain.shared
 
 import cats.Functor
 import cats.syntax.functor._
-import coop.rchain.casper.helper.HashSetCasperTestNode.Effect
 import coop.rchain.catscontrib.TaskContrib.TaskOps
+import monix.eval.Task
 import monix.execution.Scheduler
 import org.scalatest.{Assertion, Matchers}
 
@@ -19,6 +19,6 @@ object scalatestcontrib extends Matchers {
       )
   }
 
-  def effectTest[T](f: Effect[T])(implicit scheduler: Scheduler): T =
+  def effectTest[T](f: Task[T])(implicit scheduler: Scheduler): T =
     f.unsafeRunSync(scheduler)
 }
